pyproject.toml
[tool.poetry]
name = "quelfilm-chatbot-app"
version = "0.1.0"
description = "webapp for choose a movie"
authors = ["thomas.marquis.dev <thomas.marquis.dev@gmail.com>"]
include = ["templates/*.html", "resources/training.md"]
packages = [
    {include = "quelfilm"}
]

[tool.poetry.scripts]
start = 'quelfilm.app:start'

[tool.poetry.dependencies]
python = "^3.7"
nlp-tools-py-lib = "0.1.1"
flask = "^1.1.1"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
